requirements.txt
flask
pillow
pytesseract
requests
